add KD tree to gui, improve DS

+++ src/main/java/model/Point.java
@@ -5,7 +5,7 @@ import java.util.Objects;
 public record Point(double x, double y) {
     @Override
     public String toString() {
-        return "(" + (int) x + ", " + (int) y + ")";
+        return "(" + x + ", " + y + ")";
     }
 
     @Override
